requirements.txt
git+https://huggingface.co/hexgrad/Kokoro-82M
phonemizer>=2.2.1
torch>=1.10.0
numpy>=1.21.0
transformers>=4.0.0
pandas>=1.3.0
scipy>=1.7.0
beautifulsoup4>=4.10.0
requests>=2.26.0
python-dotenv>=0.19.0
openai>=0.27.0
munch>=2.5.0
